Refactor hang dump process lifetime handler
Split the handler into focused partial files for lifecycle, disposal, process-tree diagnostics, dump orchestration, and dump writing.
